Move files around to try and create some consistency and order
