+++ src/modals/templates/occupant.js
@@ -1,6 +1,6 @@
 import { html } from "lit-html";
 import { modal_close_button, modal_header_close_button } from "./buttons.js"
-import { renderAvatar } from '../../templates/directives/avatar';
+import { renderAvatar } from 'shared/directives/avatar';
 
 
 export default (o) => {
