+++ src/plugins/bookmark-views/form.js
@@ -1,5 +1,5 @@
 import tpl_muc_bookmark_form from './templates/form.js';
-import { CustomElement } from 'components/element';
+import { CustomElement } from 'shared/components/element';
 import { _converse, api } from "@converse/headless/core";
