+++ src/plugins/controlbox/toggle.js
@@ -1,5 +1,5 @@
 import tpl_controlbox_toggle from "./templates/toggle.js";
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { _converse, api } from "@converse/headless/core";
 import { showControlBox } from './utils.js';
