+++ src/plugins/muc-views/adhoc-commands.js
@@ -1,7 +1,7 @@
 import 'shared/autocomplete/index.js';
 import log from "@converse/headless/log";
 import tpl_adhoc from './templates/ad-hoc.js';
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { __ } from 'i18n';
 import { api, converse } from "@converse/headless/core";
 import { fetchCommandForm } from './utils.js';
