+++ src/plugins/muc-views/disconnected.js
@@ -1,5 +1,5 @@
 import tpl_muc_disconnect from './templates/muc-disconnect.js';
-import { CustomElement } from 'components/element';
+import { CustomElement } from 'shared/components/element';
 import { __ } from 'i18n';
 import { _converse, api } from "@converse/headless/core";
