+++ src/plugins/rootview/root.js
@@ -1,6 +1,6 @@
 import tpl_root from "./templates/root.js";
 import { api } from '@converse/headless/core';
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 
 
 /**
