+++ src/shared/autocomplete/component.js
@@ -1,5 +1,5 @@
 import AutoComplete from './autocomplete.js';
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { FILTER_CONTAINS, FILTER_STARTSWITH } from './utils.js';
 import { api } from '@converse/headless/core';
 import { html } from 'lit-element';
