+++ src/shared/chat/emoji-picker-content.js
@@ -1,4 +1,4 @@
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { _converse, converse, api } from "@converse/headless/core";
 import { html } from "lit-element";
 import { tpl_all_emojis, tpl_search_results } from "./templates/emoji-picker.js";
