+++ src/shared/chat/emoji-picker.js
@@ -1,8 +1,8 @@
 import "./emoji-picker-content.js";
 import DOMNavigator from "shared/dom-navigator";
 import debounce from 'lodash/debounce';
-import { BaseDropdown } from "components/dropdown.js";
-import { CustomElement } from 'components/element.js';
+import { BaseDropdown } from "shared/components/dropdown.js";
+import { CustomElement } from 'shared/components/element.js';
 import { __ } from 'i18n';
 import { _converse, api, converse } from "@converse/headless/core";
 import { html } from "lit-element";
