+++ src/shared/chat/message-body.js
@@ -1,7 +1,7 @@
 import 'shared/registry.js';
 import ImageModal from 'modals/image.js';
 import renderRichText from 'shared/directives/rich-text.js';
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { api } from "@converse/headless/core";
