+++ src/shared/chat/message-history.js
@@ -1,7 +1,7 @@
 import "./message";
 import dayjs from 'dayjs';
 import tpl_new_day from "templates/new_day.js";
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { _converse, api } from "@converse/headless/core";
 import { html } from 'lit-element';
 import { repeat } from 'lit-html/directives/repeat.js';
