+++ src/shared/chat/templates/message.js
@@ -1,7 +1,7 @@
 import 'shared/chat/unfurl';
 import { __ } from 'i18n';
 import { html } from "lit-html";
-import { renderAvatar } from 'templates/directives/avatar';
+import { renderAvatar } from 'shared/directives/avatar';
 
 
 export default (o) => {
