+++ src/shared/chat/unfurl.js
@@ -1,4 +1,4 @@
-import { CustomElement } from 'components/element.js';
+import { CustomElement } from 'shared/components/element.js';
 import { _converse, api } from "@converse/headless/core";
 import tpl_unfurl from './templates/unfurl.js';
