+++ src/shared/components/brand-logo.js
