+++ src/shared/components/font-awesome.js
