+++ src/shared/components/icons.js
