+++ src/shared/components/image-picker.js
@@ -1,7 +1,7 @@
 import { CustomElement } from './element.js';
-import { __ } from '../i18n';
+import { __ } from 'i18n';
 import { html } from 'lit-element';
-import { renderAvatar } from "../templates/directives/avatar.js";
+import { renderAvatar } from "shared/directives/avatar.js";
 import { api } from "@converse/headless/core";
 
 const i18n_alt_avatar = __('Your avatar image');
