+++ src/shared/directives/image.js
