+++ src/shared/directives/rich-text.js
@@ -1,4 +1,4 @@
-import { RichText } from 'shared/message/text.js';
+import { RichText } from 'shared/rich-text.js';
 import { directive, html } from "lit-html";
 import { until } from 'lit-html/directives/until.js';
