+++ src/shared/templates/avatar.js
