+++ src/templates/image.js
@@ -1,4 +1,4 @@
 import { html } from "lit-html";
-import { renderImage } from "./directives/image.js";
+import { renderImage } from "shared/directives/image.js";
 
 export default (o) => html`${renderImage(o.url, o.url, o.onLoad, o.onClick)}`;
